\name{SensMixedUI}
\alias{SensMixedUI}


\title{Graphical User interface for the SensMixed package}

\description{
launches a shiny application that provides the  graphical user interface (GUI) for the functions contained in the 
SensMixed package. Application also includes such
crucial functionalities as importing the data in different formats, presenting
results in tables and plots as well as saving them.
}
\usage{
SensMixedUI()
}

\author{
Alexandra Kuznetsova, Per Bruun Brockhoff, Rune Haubo Bojesen Christensen
}

\seealso{\link{sensmixed}, \link{conjoint}}

\examples{

\dontrun{
library(SensMixed)
SensMixedUI()
}

}